Upgrade to zend-expressive-router 57ebf52
Picks up a changer in the `RouteResult::fromRouteFailure()` signature,
as well as in what `RouteResult::getAllowedMethods()` returns when any
method is allowed.

Updates the `FastRouteRouter` class to always pass an argument to
`RouteResult::fromRouteFailure()`, passing `Route::HTTP_METHOD_ANY`
where it previously passed no argument, as these cases were cases where
the failure was not due to the HTTP method.