Ensures we can marshal a 405 result
When no routes match the combination of path and method, but we have a
successful match, then we need to get the list of allowed methods. We do
that by looping over all routes and looking for those that match the
path; when they do, we pull their allowed methods and aggregate them.